requirements.txt
# Install with pip install -r requirements.txt
requests==2.8.0
tornado==4.2.1
tornado-utils-3==1.0.0

